requirement.txt
pandas
numpy
scikit-learn
nltk
streamlit
PyPDF2
python-docx
joblib
fastapi
uvicorn
